code refactor
common bean of RestTemplate
updated scope for additionsl Gson method
separated consetns by creatong separate method for code blocks
create common currenci data kt file
