--- src/main/kotlin/com/khomishchak/cryptopricingservice/model/internall/exchanger/UsedToken.kt
@@ -1,4 +1,0 @@
-package com.khomishchak.cryptopricingservice.model.internall.exchanger
-
-data class UsedToken(val code: String,
-                     val currencies: String)
