+++ src/main/kotlin/com/khomishchak/cryptopricingservice/model/internall/exchanger/UsedTokens.kt
@@ -1,3 +1,6 @@
 package com.khomishchak.cryptopricingservice.model.internall.exchanger
 
 data class UsedTokens(val records: List<UsedToken>)
+
+data class UsedToken(val code: String,
+                     val currencies: String)
