+++ src/main/kotlin/com/khomishchak/cryptopricingservice/service/integration/exchangers/white_bit/WhiteBitCurrencyData.kt
@@ -1,0 +1,4 @@
+package com.khomishchak.cryptopricingservice.service.integration.exchangers.white_bit
+
+val stablecoins = setOf("UAH", "USD", "PLN")
+val shouldBeMappedTickets = mapOf("USDT" to "USD", "WBT-HOLD" to "WBT")
