+++ src/main/kotlin/com/khomishchak/cryptopricingservice/utility/GsonExtensions.kt
@@ -1,0 +1,7 @@
+package com.khomishchak.cryptopricingservice.utility
+
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+
+inline fun <reified T> Gson.mapJsonResp(json: String): T =
+        this.fromJson(json, object : TypeToken<T>() {}.type)
